Adjust blour radious
We want to see always the same area not blured.
To do that we need to calculate the radious for blur each time the zoom change.

The secound adjustment is realated to mobile devices.
We want to adjust the blur also to the modile screens.
Since they have different DPR we considered it in our calculations.